backend/requirements.txt
# backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0

Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Werkzeug==3.1.3

Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10

stripe==10.12.0
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.8.2

# security headers
Flask-Talisman==1.1.0
